player: change --cache-pause-initial to number of seconds
This allows the user to control the size of the initial buffer
separately from any buffering that needs to happen during playback.

+++ options/options.h
@@ -266,7 +266,7 @@ typedef struct MPOpts {
     char *sub_demuxer_name;
 
     int cache_pause;
-    int cache_pause_initial;
+    float cache_pause_initial;
     float cache_pause_wait;
 
     struct image_writer_opts *screenshot_image_opts;
